move experimental joint rupture code to separate namespace
